BUILD: Transformando pinos em uint8_t

+++ include/eletronica/pinagem.hpp
@@ -2,17 +2,17 @@
 
 // Pinagem placa NDA
 
-#define PINO_LED 0
+#define PINO_LED (uint8_t)0
 
-#define PINO_MOD_START 1
+#define PINO_MOD_START (uint8_t)1
 
-#define PINO_IN1_ME 3
-#define PINO_IN2_ME 4
-#define PINO_IN1_MD 5
-#define PINO_IN2_MD 6
+#define PINO_IN1_ME (uint8_t)3
+#define PINO_IN2_ME (uint8_t)4
+#define PINO_IN1_MD (uint8_t)5
+#define PINO_IN2_MD (uint8_t)6
 
-#define PINO_SENSOR_LATERAL_DIR 7
-#define PINO_SENSOR_FRONTAL 10
-#define PINO_SENSOR_LATERAL_ESQ 21
+#define PINO_SENSOR_LATERAL_DIR (uint8_t)7
+#define PINO_SENSOR_FRONTAL (uint8_t)10
+#define PINO_SENSOR_LATERAL_ESQ (uint8_t)21
 
-#define PINO_SENSOR_BORDA 20
+#define PINO_SENSOR_BORDA (uint8_t)20
